Move chat to base_language_predictor
