+++ tests/unit/datarobot_drum/drum/language_predictors/python_predictor/test_python_predictor.py
@@ -71,160 +71,3 @@ class TestMLPiperConfigure:
             user_secrets_mount_path=mount_path,
             user_secrets_prefix=prefix,
         )
-
-
-class TestChat:
-    @pytest.fixture
-    def mock_mlops(self):
-        with patch(
-            "datarobot_drum.drum.language_predictors.python_predictor.python_predictor.MLOps"
-        ) as mock:
-            mlops_instance = Mock()
-            mock.return_value = mlops_instance
-            yield mlops_instance
-
-    @pytest.fixture
-    def python_predictor(self, chat_python_model_adapter, mock_mlops):
-        predictor = PythonPredictor()
-        params = {
-            "target_type": TargetType.TEXT_GENERATION,
-            "__custom_model_path__": "/non-existing-path-to-avoid-loading-unwanted-artifacts",
-        }
-
-        with patch.dict(os.environ, {"TARGET_NAME": "target"}):
-            predictor.mlpiper_configure(params)
-
-        yield predictor
-
-    @pytest.mark.parametrize("response", ["Wrong response", None])
-    @pytest.mark.parametrize("stream", [False, True])
-    def test_hook_wrong_response_type(
-        self, python_predictor, chat_python_model_adapter, stream, response
-    ):
-        def chat_hook(completion_request, model):
-            return response
-
-        chat_python_model_adapter.chat_hook = chat_hook
-
-        with pytest.raises(
-            Exception,
-            match=r"Expected response to be ChatCompletion or Iterable\[ChatCompletionChunk\].*",
-        ):
-            response = python_predictor.chat(
-                {
-                    "model": "any",
-                    "messages": [
-                        {"role": "system", "content": "You are a helpful assistant."},
-                        {"role": "user", "content": "Hello!"},
-                    ],
-                    "stream": stream,
-                }
-            )
-
-            # Streaming response needs to be consumed for anything to happen
-            if stream:
-                [chunk for chunk in response]
-
-    def test_mlops_init(self, python_predictor, mock_mlops):
-        mock_mlops.set_channel_config.called_once_with("spooler_type=API")
-
-        mock_mlops.init.assert_called_once()
-
-    @pytest.mark.parametrize("stream", [False, True])
-    def test_chat_with_mlops(self, python_predictor, chat_python_model_adapter, mock_mlops, stream):
-        def chat_hook(completion_request, model):
-            return (
-                create_completion_chunks(["How", " are", " you"])
-                if stream
-                else create_completion("How are you")
-            )
-
-        chat_python_model_adapter.chat_hook = chat_hook
-
-        response = python_predictor.chat(
-            {
-                "model": "any",
-                "messages": [
-                    {"role": "system", "content": "You are a helpful assistant."},
-                    {"role": "user", "content": "Hello!"},
-                ],
-                "stream": stream,
-            }
-        )
-        if stream:
-            # Streaming response needs to be consumed for anything to happen
-            [chunk for chunk in response]
-
-        mock_mlops.report_deployment_stats.assert_called_once_with(
-            num_predictions=1, execution_time_ms=ANY
-        )
-
-        mock_mlops.report_predictions_data.assert_called_once_with(
-            ANY,
-            ["How are you"],
-            association_ids=ANY,
-            skip_drift_tracking=True,
-            skip_accuracy_tracking=True,
-        )
-        # Compare features dataframe separately as this doesn't play nice with assert_called
-        assert mock_mlops.report_predictions_data.call_args.args[0]["prompt"].values[0] == "Hello!"
-
-    @pytest.mark.parametrize("stream", [False, True])
-    def test_failing_hook_with_mlops(
-        self, python_predictor, chat_python_model_adapter, mock_mlops, stream
-    ):
-        def chat_hook(completion_request, model):
-            raise BadRequest("Error")
-
-        chat_python_model_adapter.chat_hook = chat_hook
-
-        with pytest.raises(BadRequest):
-            response = python_predictor.chat(
-                {
-                    "model": "any",
-                    "messages": [
-                        {"role": "system", "content": "You are a helpful assistant."},
-                        {"role": "user", "content": "Hello!"},
-                    ],
-                    "stream": stream,
-                }
-            )
-
-        mock_mlops.report_deployment_stats.assert_called_once_with(
-            num_predictions=0, execution_time_ms=ANY
-        )
-        mock_mlops.report_predictions_data.assert_not_called()
-
-    def test_failing_in_middle_of_stream(
-        self, python_predictor, chat_python_model_adapter, mock_mlops
-    ):
-        def chat_hook(completion_request, model):
-            def generator():
-                for chunk in create_completion_chunks(["Chunk1", "Chunk2"]):
-                    yield chunk
-
-                raise BadRequest("Error")
-
-            return generator()
-
-        chat_python_model_adapter.chat_hook = chat_hook
-
-        with pytest.raises(BadRequest):
-            response = python_predictor.chat(
-                {
-                    "model": "any",
-                    "messages": [
-                        {"role": "system", "content": "You are a helpful assistant."},
-                        {"role": "user", "content": "Hello!"},
-                    ],
-                    "stream": True,
-                }
-            )
-
-            # Streaming response needs to be consumed for anything to happen
-            [chunk for chunk in response]
-
-        mock_mlops.report_deployment_stats.assert_called_once_with(
-            num_predictions=0, execution_time_ms=ANY
-        )
-        mock_mlops.report_predictions_data.assert_not_called()
